change skills entity : rating is now a float

+++ api/src/Entity/Skill.php
@@ -39,7 +39,7 @@ class Skill
     private $skills;
 
     /**
-     * @ORM\Column(type="smallint")
+     * @ORM\Column(type="float")
      */
     private $rating;
 
@@ -112,12 +112,12 @@ class Skill
         return $this;
     }
 
-    public function getRating(): ?int
+    public function getRating(): ?float
     {
         return $this->rating;
     }
 
-    public function setRating(int $rating): self
+    public function setRating(float $rating): self
     {
         $this->rating = $rating;
